Added main.py as entrypoint instead of run.py. Added typehints to create_bot.py,
